requirements.txt
flask==2.2.5
pandas==1.5.3
psycopg2-binary==2.9.6
kafka-python==2.0.2
xgboost==3.0.0
joblib==1.3.2
numpy==1.24.3
scikit-learn==1.4.2
matplotlib==3.7.2
optuna==3.3.0
flask-socketio==5.3.4
requests==2.28.1
